feat(p0149): migrate DI to feature-set extension methods
Replaces the three monolithic ServiceCollectionExtensions.cs files
(Application 442 / Infrastructure 194 / Server 242 lines) with one
public AddXxx() extension method per feature-set, each living next to
the services it registers. Composition roots become a flat list of
AddXxx() calls — removing one call removes one feature-set; the
program still compiles. Matches the RHS.Common.Commands pattern.

Before / after composition-root sizes:
- Application: 475 → 37 lines (10 AddXxx calls)
- Infrastructure: 52 → 37 lines (9 AddXxx calls)
- Server: 72 → 14 lines (marker; Program.cs chains the AddXxx calls)

Application feature-sets:
- AddCommandExecutor, AddPromptCatalog, AddConfigurationValidation
- AddPipelineHandlers, AddSkillRunHandlers, AddSkillRoundInfrastructure
- AddContextBuilders, AddPipelineExecution, AddLoopRuntime
- AddSwaggerCompression, AddWebhookCommentIntent

Infrastructure feature-sets:
- AddTicketProviders, AddSourceProviders, AddAgentProviders
- AddOutputStrategies, AddContainerRunners, AddSecurityScanners
- AddDialogueTransport, AddProjectMeta

Server feature-sets:
- AddRedis, AddCoreDispatcherServices, AddServerCompositionOverrides
- AddLongRunningServices, AddIntentEngine, AddIntentHandlers
- AddWebhookHandlers, AddSlackAdapter, AddTeamsAdapter
- AddSandbox + 3 sandbox/orchestrator config methods
- AddJobSpawnerOptions

Drops the seven dormant p0147h partials from Application/Extensions/
that were never wired into AddAgentSmithCommands. Triple-registered
concept-publishing handlers (CheckoutSource, TryCheckoutSource,
PipelineNameInitializer, BootstrapCheck, PublishProjectLanguage) are
collapsed via a single generic helper.

Grep audit: `grep -rn 'ServiceCollectionExtensions\.[A-Z]' src/` → 0
references; the dormant partials are gone.

Validation:
- dotnet build: 0 errors, 0 warnings
- dotnet test: 1786 + 66 = 1852 passing (parity with main)
- CLI --help boots; api-scan / security-scan / fix / feature --help all green
- docker compose -f deploy/docker-compose.example.yml config -q → exit 0
