--- src/AgentSmith.Application/Extensions/ServiceCollectionExtensions.SecurityHandlers.cs
@@ -1,35 +1,0 @@
-using AgentSmith.Application.Models;
-using AgentSmith.Application.Services;
-using AgentSmith.Application.Services.Handlers;
-using AgentSmith.Contracts.Commands;
-using AgentSmith.Contracts.Services;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace AgentSmith.Application;
-
-public static partial class ServiceCollectionExtensions
-{
-    // Security + API-security scanner handlers. Covers the static / git-history /
-    // dependency-audit chain plus the Swagger/Nuclei/Spectral/ZAP API-side, the
-    // compression + trend handlers, and the SpawnFix auto-remediation step.
-    private static void AddSecurityHandlers(IServiceCollection services)
-    {
-        services.AddTransient<ICommandHandler<SecuritySkillRoundContext>, SecuritySkillRoundHandler>();
-        services.AddTransient<ICommandHandler<LoadSwaggerContext>, LoadSwaggerHandler>();
-        services.AddTransient<ICommandHandler<SpawnNucleiContext>, SpawnNucleiHandler>();
-        services.AddTransient<ICommandHandler<SpawnSpectralContext>, SpawnSpectralHandler>();
-        services.AddTransient<ICommandHandler<SpawnZapContext>, SpawnZapHandler>();
-        services.AddTransient<ICommandHandler<ApiSecuritySkillRoundContext>, ApiSkillRoundHandler>();
-        services.AddTransient<ICommandHandler<CompileFindingsContext>, CompileFindingsHandler>();
-        services.AddTransient<ICommandHandler<DeliverFindingsContext>, DeliverFindingsHandler>();
-        services.AddTransient<ICommandHandler<StaticPatternScanContext>, StaticPatternScanHandler>();
-        services.AddTransient<ICommandHandler<GitHistoryScanContext>, GitHistoryScanHandler>();
-        services.AddTransient<ICommandHandler<DependencyAuditContext>, DependencyAuditHandler>();
-        services.AddTransient<ICommandHandler<CompressSecurityFindingsContext>, CompressSecurityFindingsHandler>();
-        services.AddTransient<ICommandHandler<CompressApiScanFindingsContext>, CompressApiScanFindingsHandler>();
-        services.AddTransient<ICommandHandler<SecurityTrendContext>, SecurityTrendHandler>();
-        services.AddTransient<ICommandHandler<SecuritySnapshotWriteContext>, SecuritySnapshotWriter>();
-        services.AddTransient<ICommandHandler<SpawnFixContext>, SpawnFixHandler>();
-        services.AddSingleton<HttpProbeRunner>();
-    }
-}
