+++ src/AgentSmith.Application/Services/Pipeline/PipelineExecutionExtensions.cs
@@ -1,13 +1,13 @@
 using AgentSmith.Application.Models;
 using AgentSmith.Application.PipelineDataFlows;
-using AgentSmith.Application.Services;
 using AgentSmith.Application.Services.Builders;
 using AgentSmith.Application.Services.Lifecycle;
 using AgentSmith.Application.Services.Orchestrator;
-using AgentSmith.Application.Services.Pipeline;
+using AgentSmith.Application.Services.Prompts;
 using AgentSmith.Application.Services.Sandbox;
+using AgentSmith.Application.Services.Spawning;
 using AgentSmith.Application.Services.Tools;
-using AgentSmith.Contracts.Commands;
+using AgentSmith.Application.Services.Triggers;
 using AgentSmith.Contracts.Models.Configuration;
 using AgentSmith.Contracts.Pipeline;
 using AgentSmith.Contracts.Sandbox;
@@ -15,15 +15,18 @@ using AgentSmith.Contracts.Services;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
-namespace AgentSmith.Application;
+namespace AgentSmith.Application.Services.Pipeline;
 
-public static partial class ServiceCollectionExtensions
+/// <summary>
+/// Pipeline-execution feature-set: executor (and the legacy fallback behind
+/// PIPELINE_EXECUTOR_USE_LEGACY), step/error/sandbox collaborators, data-flow
+/// resolver + per-preset declarations, sandbox + image resolvers, lifecycle
+/// coordinator default, prompt builder, sandbox file-reader factory,
+/// project resolver + spawn use-case, and the IntentParser binding.
+/// </summary>
+public static class PipelineExecutionExtensions
 {
-    // Pipeline orchestration core: executor, intent parsing, sandbox + image resolvers,
-    // per-preset IPhaseDataFlow registrations, project + spawn use-cases, and the
-    // ToolKit policy. p0128c registers each IPhaseDataFlow as a singleton so the
-    // resolver builds an O(1) name→declaration index at startup.
-    private static void AddPipelineRuntime(IServiceCollection services)
+    public static IServiceCollection AddPipelineExecution(this IServiceCollection services)
     {
         services.AddTransient<IIntentParser>(sp => new LlmIntentParser(
             sp.GetRequiredService<IChatClientFactory>(),
@@ -31,7 +34,15 @@ public static partial class ServiceCollectionExtensions
             new AgentConfig { Type = "claude" },
             sp.GetRequiredService<ILogger<LlmIntentParser>>()));
         services.AddTransient<ICommandContextFactory, CommandContextFactory>();
-        services.AddTransient<IPipelineExecutor, PipelineExecutor>();
+        services.AddTransient<IPipelineStepRunner, PipelineStepRunner>();
+        services.AddTransient<IPipelineErrorHandler, PipelineErrorHandler>();
+        services.AddTransient<IPipelineSandboxCoordinator, PipelineSandboxCoordinator>();
+        services.AddTransient<PipelineExecutor>();
+        services.AddTransient<PipelineExecutorLegacy>();
+        services.AddTransient<IPipelineExecutor>(sp =>
+            UseLegacyExecutor()
+                ? sp.GetRequiredService<PipelineExecutorLegacy>()
+                : sp.GetRequiredService<PipelineExecutor>());
         services.AddSingleton<IPhaseDataFlow, FixBugDataFlow>();
         services.AddSingleton<IPhaseDataFlow, FixNoTestDataFlow>();
         services.AddSingleton<IPhaseDataFlow, AddFeatureDataFlow>();
@@ -50,25 +61,22 @@ public static partial class ServiceCollectionExtensions
         services.AddTransient<ISandboxLanguageResolver, SandboxLanguageResolver>();
         services.AddTransient<ISourceConfigOverrider, SourceConfigOverrider>();
         services.AddSingleton<IPipelineConfigResolver, PipelineConfigResolver>();
-        // p0140a/b: ProjectResolver is stateless; exposed as IEnvelopeProjectResolver for
-        // webhook handlers + SpawnPipelineRunsUseCase.
-        services.AddSingleton<Services.Triggers.ProjectResolver>();
+        services.AddSingleton<ProjectResolver>();
         services.AddSingleton<IEnvelopeProjectResolver>(
-            sp => sp.GetRequiredService<Services.Triggers.ProjectResolver>());
-        // p0140b: SpawnPipelineRunsUseCase builds ClaimRequests from a webhook envelope.
-        // Depends on ITicketClaimService (Server-only) so this service resolves only inside
-        // the Server composition; CLI graph doesn't use it.
-        services.AddTransient<ISpawnPipelineRunsUseCase, Services.Spawning.SpawnPipelineRunsUseCase>();
+            sp => sp.GetRequiredService<ProjectResolver>());
+        services.AddTransient<ISpawnPipelineRunsUseCase, SpawnPipelineRunsUseCase>();
         services.AddTransient<ExecutePipelineUseCase>();
-        // ITicketClaimService moved to Server.AddCoreDispatcherServices in p0109a. Application's
-        // PipelineExecutor delegates lifecycle wrapping to IPipelineLifecycleCoordinator (NoOp
-        // by default; Server overrides).
         services.AddSingleton<IPipelineLifecycleCoordinator, NoOpPipelineLifecycleCoordinator>();
-        services.AddSingleton<Services.Prompts.AgentPromptBuilder>();
+        services.AddSingleton<AgentPromptBuilder>();
         services.AddSingleton<ISandboxFileReaderFactory, SandboxFileReaderFactory>();
-        // p0145: IToolHost instances are NOT DI-registered — each host carries per-pipeline-run
-        // state. ToolKit is stateless (only the policy), so singleton is fine.
         services.AddSingleton<IPipelineToolPolicy, AllHostsActivePolicy>();
         services.AddSingleton<IToolKit, ToolKit>();
+        return services;
+    }
+
+    private static bool UseLegacyExecutor()
+    {
+        var raw = Environment.GetEnvironmentVariable("PIPELINE_EXECUTOR_USE_LEGACY");
+        return raw is "1" or "true" or "TRUE";
     }
 }
