+++ src/AgentSmith.Infrastructure/Extensions/ServiceCollectionExtensions.cs
@@ -1,52 +1,37 @@
-using AgentSmith.Contracts.Activation;
-using AgentSmith.Contracts.Commands;
-using AgentSmith.Contracts.Models;
-using AgentSmith.Contracts.Models.Skills;
-using AgentSmith.Contracts.Services;
 using AgentSmith.Infrastructure.Core;
-using AgentSmith.Infrastructure.Services.Activation;
+using AgentSmith.Infrastructure.Services;
+using AgentSmith.Infrastructure.Services.Containers;
+using AgentSmith.Infrastructure.Services.Dialogue;
+using AgentSmith.Infrastructure.Services.Output;
+using AgentSmith.Infrastructure.Services.Providers.Agent;
+using AgentSmith.Infrastructure.Services.Providers.Source;
+using AgentSmith.Infrastructure.Services.Providers.Tickets;
+using AgentSmith.Infrastructure.Services.Security;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace AgentSmith.Infrastructure;
 
 /// <summary>
-/// Registers all infrastructure services with the DI container. Helpers are split
-/// into partial files by subdomain (TicketProviders, SourceProviders, ChatClients,
-/// OutputStrategies, SecurityScanners, Dialogue, ConfigLoading) so each file stays
-/// under the 120-line limit. The public entry point lives here.
+/// Infrastructure composition root: a flat list of per-feature-set Add calls.
+/// Removing one call removes one feature-set; the program still compiles. Each
+/// feature-set's AddXxx() lives next to the services it registers.
 /// </summary>
-public static partial class ServiceCollectionExtensions
+public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddAgentSmithInfrastructure(this IServiceCollection services)
     {
         services.AddAgentSmithCore();
-        // p0137b: AddHttpClient() is called once at the composition root (Server's
-        // Program.cs / CLI's ServiceProviderFactory). Per-feature extensions use
-        // AddHttpClient<T>() typed clients instead of the bare factory registration.
-        AddTicketProviders(services);
-        AddSourceProviders(services);
-        AddChatClients(services);
-        AddOutputStrategies(services);
-        AddSecurityScanners(services);
-        AddDialogue(services);
-        // Project-meta resolution under target SourcePath (p0104)
-        services.AddSingleton<IProjectMetaResolver, Services.ProjectMetaResolver>();
-        services.AddSingleton<IProjectBriefBuilder, Services.ProjectBriefBuilder>();
-        services.AddSingleton<IBaselineLoader, Services.BaselineLoader>();
-        // p0125b: PipelineContextRunStateConcepts is bound to a per-pipeline PipelineContext
-        // (not a DI singleton). Register a factory so handlers (p0125c) can inject the
-        // creation surface; the factory pulls vocabulary from the context's ConceptVocabulary
-        // slot, falling back to Empty when no skills are loaded yet (test fixtures).
-        services.AddSingleton<Func<PipelineContext, IRunStateConcepts>>(_ => CreateRunStateConcepts);
+        // AddHttpClient() is called once at the composition root (Server's Program.cs
+        // / CLI's ServiceProviderFactory). Per-feature extensions use AddHttpClient<T>()
+        // typed clients instead of the bare factory registration.
+        services.AddTicketProviders();
+        services.AddSourceProviders();
+        services.AddAgentProviders();
+        services.AddOutputStrategies();
+        services.AddContainerRunners();
+        services.AddSecurityScanners();
+        services.AddDialogueTransport();
+        services.AddProjectMeta();
         return services;
-    }
-
-    private static IRunStateConcepts CreateRunStateConcepts(PipelineContext context)
-    {
-        var vocabulary = context.TryGet<ConceptVocabulary>(ContextKeys.ConceptVocabulary, out var loaded)
-            && loaded is not null
-                ? loaded
-                : ConceptVocabulary.Empty;
-        return new PipelineContextRunStateConcepts(context, vocabulary);
     }
 }
