--- src/AgentSmith.Server/Extensions/SandboxServiceCollectionExtensions.cs
@@ -1,160 +1,0 @@
-using AgentSmith.Application.Models;
-using AgentSmith.Application.Services.Orchestrator;
-using AgentSmith.Contracts.Models.Configuration;
-using AgentSmith.Contracts.Sandbox;
-using AgentSmith.Contracts.Services;
-using AgentSmith.Infrastructure.Services.Sandbox;
-using AgentSmith.Server.Services.Orchestrator;
-using AgentSmith.Server.Services.Sandbox;
-using Docker.DotNet;
-using k8s;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using StackExchange.Redis;
-
-namespace AgentSmith.Server.Extensions;
-
-internal static class SandboxServiceCollectionExtensions
-{
-    /// <summary>
-    /// Binds global sandbox-container resource defaults from configuration section "Sandbox"
-    /// (appsettings.json + environment variables with the Sandbox__ prefix). Per-project
-    /// blocks in agentsmith.yml under projects.&lt;name&gt;.sandbox.resources override these.
-    /// </summary>
-    internal static IServiceCollection AddSandboxOptions(
-        this IServiceCollection services, IConfiguration configuration)
-    {
-        services.Configure<SandboxOptions>(configuration.GetSection("Sandbox"));
-        return services;
-    }
-
-    /// <summary>
-    /// Registers <see cref="IOptions{TOptions}"/> of <see cref="SandboxGlobalConfig"/>
-    /// pulled from the loaded agentsmith.yml top-level <c>sandbox:</c> block.
-    /// Consumed by <see cref="AgentSmith.Application.Services.Sandbox.AgentImageResolver"/>
-    /// (and any future agentsmith.yml-driven sandbox setting).
-    /// </summary>
-    internal static IServiceCollection AddSandboxGlobalConfig(this IServiceCollection services)
-    {
-        services.AddSingleton<IOptions<SandboxGlobalConfig>>(sp =>
-        {
-            var loader = sp.GetRequiredService<IConfigurationLoader>();
-            var context = sp.GetRequiredService<ServerContext>();
-            return Options.Create(loader.LoadConfig(context.ConfigPath).Sandbox);
-        });
-        return services;
-    }
-
-    /// <summary>
-    /// Registers <see cref="IOptions{TOptions}"/> of <see cref="OrchestratorGlobalConfig"/>
-    /// pulled from the loaded agentsmith.yml top-level <c>orchestrator:</c> block.
-    /// Consumed by <see cref="AgentSmith.Application.Services.Orchestrator.OrchestratorImageResolver"/>.
-    /// Also registers <see cref="IOrchestratorResourceResolver"/> backed by
-    /// <see cref="OrchestratorResourceResolver"/> (Server impl — accesses Server-resident
-    /// <c>JobSpawnerOptions.Resources</c> as the global-default surface).
-    /// </summary>
-    internal static IServiceCollection AddOrchestratorGlobalConfig(this IServiceCollection services)
-    {
-        services.AddSingleton<IOptions<OrchestratorGlobalConfig>>(sp =>
-        {
-            var loader = sp.GetRequiredService<IConfigurationLoader>();
-            var context = sp.GetRequiredService<ServerContext>();
-            return Options.Create(loader.LoadConfig(context.ConfigPath).Orchestrator);
-        });
-        services.AddSingleton<IOrchestratorResourceResolver, OrchestratorResourceResolver>();
-        return services;
-    }
-
-    /// <summary>
-    /// Auto-detects the sandbox backend at startup. Order:
-    ///   SANDBOX_TYPE explicit > KUBERNETES_SERVICE_HOST > /var/run/docker.sock > inprocess fallback.
-    /// </summary>
-    internal static IServiceCollection AddSandbox(this IServiceCollection services)
-    {
-        var backend = ResolveBackend();
-        switch (backend)
-        {
-            case SandboxBackend.Kubernetes:
-                RegisterKubernetes(services);
-                break;
-            case SandboxBackend.Docker:
-                RegisterDocker(services);
-                break;
-            case SandboxBackend.InProcess:
-                services.AddSingleton<ISandboxFactory, InProcessSandboxFactory>();
-                break;
-        }
-        services.AddSingleton(new SandboxBackendInfo(backend));
-        return services;
-    }
-
-    private static SandboxBackend ResolveBackend()
-    {
-        var explicitType = Environment.GetEnvironmentVariable("SANDBOX_TYPE");
-        if (string.Equals(explicitType, "kubernetes", StringComparison.OrdinalIgnoreCase))
-            return SandboxBackend.Kubernetes;
-        if (string.Equals(explicitType, "docker", StringComparison.OrdinalIgnoreCase))
-            return SandboxBackend.Docker;
-        if (string.Equals(explicitType, "inprocess", StringComparison.OrdinalIgnoreCase))
-            return SandboxBackend.InProcess;
-        if (!string.IsNullOrEmpty(Environment.GetEnvironmentVariable("KUBERNETES_SERVICE_HOST")))
-            return SandboxBackend.Kubernetes;
-        if (File.Exists("/var/run/docker.sock"))
-            return SandboxBackend.Docker;
-        return SandboxBackend.InProcess;
-    }
-
-    private static void RegisterKubernetes(IServiceCollection services)
-    {
-        services.AddSingleton<IKubernetes>(_ =>
-        {
-            var config = KubernetesClientConfiguration.InClusterConfig();
-            return new Kubernetes(config);
-        });
-        services.AddSingleton<PodSpecBuilder>();
-        services.AddSingleton(new KubernetesSandboxOptions
-        {
-            Namespace = Environment.GetEnvironmentVariable("K8S_NAMESPACE") ?? "default",
-            RedisUrl = Environment.GetEnvironmentVariable("REDIS_URL") ?? "redis:6379"
-        });
-        services.AddSingleton<ISandboxFactory>(sp => new KubernetesSandboxFactory(
-            sp.GetRequiredService<IKubernetes>(),
-            sp.GetRequiredService<IConnectionMultiplexer>(),
-            sp.GetRequiredService<PodSpecBuilder>(),
-            sp.GetRequiredService<KubernetesSandboxOptions>(),
-            sp.GetRequiredService<ILoggerFactory>()));
-    }
-
-    private static void RegisterDocker(IServiceCollection services)
-    {
-        services.AddSingleton(new DockerSandboxOptions
-        {
-            RedisUrl = Environment.GetEnvironmentVariable("REDIS_URL") ?? "redis:6379",
-            DockerSocketUri = Environment.GetEnvironmentVariable("DOCKER_HOST") ?? "unix:///var/run/docker.sock",
-            Network = Environment.GetEnvironmentVariable("DOCKER_NETWORK") ?? ""
-        });
-        services.AddSingleton<IDockerClient>(sp =>
-        {
-            var opts = sp.GetRequiredService<DockerSandboxOptions>();
-            return new DockerClientConfiguration(new Uri(opts.DockerSocketUri)).CreateClient();
-        });
-        services.AddSingleton<DockerContainerSpecBuilder>();
-        services.AddSingleton<ISandboxFactory>(sp => new DockerSandboxFactory(
-            sp.GetRequiredService<IDockerClient>(),
-            sp.GetRequiredService<IConnectionMultiplexer>(),
-            sp.GetRequiredService<DockerContainerSpecBuilder>(),
-            sp.GetRequiredService<DockerSandboxOptions>(),
-            sp.GetRequiredService<ILoggerFactory>()));
-    }
-}
-
-internal enum SandboxBackend
-{
-    InProcess,
-    Kubernetes,
-    Docker
-}
-
-internal sealed record SandboxBackendInfo(SandboxBackend Backend);
