+++ src/AgentSmith.Server/Services/Hosting/PollingExtensions.cs
@@ -5,11 +5,13 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace AgentSmith.Server.Extensions;
 
-internal static partial class ServiceCollectionExtensions
+/// <summary>
+/// Long-running hosted services: queue consumer, housekeeping leader, poller
+/// leader, and the Redis connection health monitor. Each one is published as
+/// ISubsystemHealth so /health endpoints expose the per-subsystem status.
+/// </summary>
+internal static class PollingExtensions
 {
-    // Long-running hosted services: queue consumer, housekeeping leader, poller leader,
-    // and the Redis connection health monitor. Each one is published as ISubsystemHealth
-    // so /health endpoints expose the per-subsystem status.
     internal static IServiceCollection AddLongRunningServices(this IServiceCollection services)
     {
         services.AddSingleton<QueueConsumerHostedService>();
